chore: 去除目前所有构建警告项
Description: 规范用法,规避老用法

Log: 去除目前所有构建警告项

link:

+++ src/djvuControl/SheetBrowserDJVU.cpp
@@ -272,15 +272,15 @@ void SheetBrowserDJVU::deform(DocOperation &operation)
     case Dr::RotateBy0:
     case Dr::RotateBy180:
         if (Dr::FitToPageWidthMode == operation.scaleMode)
-            operation.scaleFactor = ((double)this->width() - 25.0) / (double) m_maxWidth / (Dr::TwoPagesMode == operation.layoutMode ? 2 : 1);
+            operation.scaleFactor = static_cast<double>(this->width() - 25.0) / static_cast<double>(m_maxWidth) / (Dr::TwoPagesMode == operation.layoutMode ? 2 : 1);
         else if (Dr::FitToPageHeightMode == operation.scaleMode)
-            operation.scaleFactor = (double)this->height() / (double) m_maxHeight;
+            operation.scaleFactor = static_cast<double>(this->height()) / static_cast<double>(m_maxHeight);
         else if (Dr::FitToPageDefaultMode == operation.scaleMode)
             operation.scaleFactor = 1.0 ;
         else if (Dr::FitToPageWorHMode == operation.scaleMode) {
-            qreal scaleFactor = ((double)this->width() - 25.0) / (double) m_maxWidth / (Dr::TwoPagesMode == operation.layoutMode ? 2 : 1);
+            qreal scaleFactor = static_cast<double>(this->width() - 25.0) / static_cast<double>(m_maxWidth) / (Dr::TwoPagesMode == operation.layoutMode ? 2 : 1);
             if (scaleFactor * m_maxHeight > this->height())
-                scaleFactor = (double)this->height() / (double) m_maxHeight;
+                scaleFactor = static_cast<double>(this->height()) / static_cast<double>(m_maxHeight);
             operation.scaleFactor = scaleFactor;
         } else
             operation.scaleMode = Dr::ScaleFactorMode;
@@ -288,15 +288,15 @@ void SheetBrowserDJVU::deform(DocOperation &operation)
     case Dr::RotateBy90:
     case Dr::RotateBy270:
         if (Dr::FitToPageWidthMode == operation.scaleMode)
-            operation.scaleFactor = ((double)this->width() - 25.0) / (double) m_maxHeight / (Dr::TwoPagesMode == operation.layoutMode ? 2 : 1);
+            operation.scaleFactor = static_cast<double>(this->width() - 25.0) / static_cast<double>(m_maxHeight) / (Dr::TwoPagesMode == operation.layoutMode ? 2 : 1);
         else if (Dr::FitToPageHeightMode == operation.scaleMode)
-            operation.scaleFactor = (double)this->height() / (double) m_maxWidth;
+            operation.scaleFactor = static_cast<double>(this->height()) / static_cast<double>(m_maxWidth);
         else if (Dr::FitToPageDefaultMode == operation.scaleMode)
             operation.scaleFactor = 1.0 ;
         else if (Dr::FitToPageWorHMode == operation.scaleMode) {
-            qreal scaleFactor = ((double)this->width() - 25.0) / (double) m_maxHeight / (Dr::TwoPagesMode == operation.layoutMode ? 2 : 1);
+            qreal scaleFactor = static_cast<double>(this->width() - 25.0) / static_cast<double>(m_maxHeight) / (Dr::TwoPagesMode == operation.layoutMode ? 2 : 1);
             if (scaleFactor * m_maxWidth > this->height())
-                scaleFactor = (double)this->height() / (double) m_maxWidth;
+                scaleFactor = static_cast<double>(this->height()) / static_cast<double>(m_maxWidth);
             operation.scaleFactor = scaleFactor;
         } else
             operation.scaleMode = Dr::ScaleFactorMode;
@@ -312,7 +312,7 @@ void SheetBrowserDJVU::deform(DocOperation &operation)
             m_items.at(i)->setPos(0, height);
             height += m_items.at(i)->boundingRect().height() + 5;
             if (m_items.at(i)->boundingRect().width() > width)
-                width = m_items.at(i)->boundingRect().width();
+                width = static_cast<int>(m_items.at(i)->boundingRect().width());
         }
     } else if (Dr::TwoPagesMode == operation.layoutMode) {
         for (int i = 0; i < m_items.count(); ++i) {
@@ -333,10 +333,10 @@ void SheetBrowserDJVU::deform(DocOperation &operation)
 
             if (m_items.count() > i + 1) {
                 if (m_items.at(i)->boundingRect().width() + m_items.at(i + 1)->boundingRect().width() > width)
-                    width = m_items.at(i)->boundingRect().width() + m_items.at(i + 1)->boundingRect().width();
+                    width = static_cast<int>(m_items.at(i)->boundingRect().width() + m_items.at(i + 1)->boundingRect().width());
             } else {
                 if (m_items.at(i)->boundingRect().width() > width)
-                    width = m_items.at(i)->boundingRect().width();
+                    width = static_cast<int>(m_items.at(i)->boundingRect().width());
             }
         }
     }
@@ -344,7 +344,7 @@ void SheetBrowserDJVU::deform(DocOperation &operation)
     setSceneRect(0, 0, width, height);
 
     if (page > 0 && page <= m_items.count()) {
-        verticalScrollBar()->setValue(m_items[page - 1]->pos().y() + m_items[page - 1]->boundingRect().height()*pageDiffFactor);
+        verticalScrollBar()->setValue(static_cast<int>(m_items[page - 1]->pos().y() + m_items[page - 1]->boundingRect().height()*pageDiffFactor));
     }
 }
 
@@ -470,7 +470,7 @@ void SheetBrowserDJVU::setCurrentPage(int page)
     if (page < 1 && page > allPages())
         return;
 
-    verticalScrollBar()->setValue(m_items.at(page - 1)->pos().y());
+    verticalScrollBar()->setValue(static_cast<int>(m_items.at(page - 1)->pos().y()));
 
     pageChanged(page);
 }
@@ -480,7 +480,7 @@ bool SheetBrowserDJVU::getImage(int index, QImage &image, double width, double h
     if (m_items.count() <= index)
         return false;
 
-    image = m_items.at(index)->getImage(width, height, mode);
+    image = m_items.at(index)->getImage(static_cast<int>(width), static_cast<int>(height), mode);
 
     return true;
 }
@@ -497,7 +497,7 @@ bool SheetBrowserDJVU::getImagePoint(QPoint viewPoint, double scaleFactor, QImag
     if (!item->contains(itemPoint))
         return false;
 
-    QPoint point = QPoint(itemPoint.x(), itemPoint.y());
+    QPoint point = QPoint(static_cast<int>(itemPoint.x()), static_cast<int>(itemPoint.y()));
 
     image = item->getImagePoint(scaleFactor, point);
